backend/requirements.txt
django-allauth[socialaccount]>=65.4.1
django-ninja >= 1.3.0,<2
PyYAML >= 6.0.2

